+++ src/NPP/DockingDlgInterface.h
@@ -22,7 +22,6 @@
 
 #include <assert.h>
 #include <shlwapi.h>
-#include <string>
 #include "StaticDialog.h"
 
 
@@ -33,21 +32,21 @@ public:
 	DockingDlgInterface() = default;
 	explicit DockingDlgInterface(int dlgID): _dlgID(dlgID) {}
 
-	virtual void init(HINSTANCE hInst, HWND parent) {
+	void init(HINSTANCE hInst, HWND parent) override {
 		StaticDialog::init(hInst, parent);
-		TCHAR temp[MAX_PATH];
-		::GetModuleFileName(reinterpret_cast<HMODULE>(hInst), temp, MAX_PATH);
+		wchar_t temp[MAX_PATH];
+		::GetModuleFileName(hInst, temp, MAX_PATH);
 		_moduleName = ::PathFindFileName(temp);
 	}
 
-    void create(tTbData* data, bool isRTL = false) {
+	void create(tTbData* data, bool isRTL = false) {
 		assert(data != nullptr);
 		StaticDialog::create(_dlgID, isRTL);
-		TCHAR temp[MAX_PATH];
+		wchar_t temp[MAX_PATH];
 		::GetWindowText(_hSelf, temp, MAX_PATH);
 		_pluginName = temp;
 
-        // user information
+		// user information
 		data->hClient = _hSelf;
 		data->pszName = _pluginName.c_str();
 
@@ -62,12 +61,10 @@ public:
 		::SendMessage(_hParent, NPPM_DMMUPDATEDISPINFO, 0, reinterpret_cast<LPARAM>(_hSelf));
 	}
 
-    virtual void destroy() {}
-
 	virtual void setBackgroundColor(COLORREF) {}
 	virtual void setForegroundColor(COLORREF) {}
 
-	virtual void display(bool toShow = true) const {
+	void display(bool toShow = true) const override {
 		::SendMessage(_hParent, toShow ? NPPM_DMMSHOW : NPPM_DMMHIDE, 0, reinterpret_cast<LPARAM>(_hSelf));
 	}
 
@@ -79,7 +76,7 @@ public:
 		_isClosed = toClose;
 	}
 
-	const TCHAR * getPluginFileName() const {
+	const wchar_t * getPluginFileName() const {
 		return _moduleName.c_str();
 	}
 
@@ -91,10 +88,10 @@ protected :
 	std::wstring _pluginName;
 	bool _isClosed = false;
 
-	virtual INT_PTR CALLBACK run_dlgProc(UINT message, WPARAM, LPARAM lParam) {
+	intptr_t CALLBACK run_dlgProc(UINT message, [[maybe_unused]] WPARAM wParam, LPARAM lParam) override {
 		switch (message)
 		{
-			case WM_NOTIFY: 
+			case WM_NOTIFY:
 			{
 				LPNMHDR	pnmh = reinterpret_cast<LPNMHDR>(lParam);